scripts/requirements.txt
svgwrite>=1.4.3
Pillow>=9.5.0